refactor(tests): Separate utils from test functions
